requirements.txt
Flask
Flask-SQLAlchemy
PyMySQL
mysql-connector-python